+++ Deprecated/UserIndexScreen.py
@@ -1,3 +1,7 @@
+'''
+
+This Class is deprecated
+'''
 import tkinter as tk
 from tkinter import ttk
 import sys