+++ Screens/__init__.py
@@ -3,8 +3,8 @@ from .CreateDataBaseScreen import CreateDatabaseScreen
 from .LoadDatabaseScreen import LoadDatabaseScreen
 from .RecipeIndexScreen import RecipeIndex
 from .ViewRecipeScreen import ViewRecipe
-from .CreateUserScreen import CreateUser
+# from .CreateUserScreen import CreateUser  # deprecated
 from .CreateAdminScreen import CreateAdmin
 from .LoginScreen import Login
-from .UserIndexScreen import UserIndex
+# from .UserIndexScreen import UserIndex  # deprecated
 from .MainMenuScreen import MainMenu